# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "DirectOriginKind": {
        "enum": [
          "purchased",
          "gift_exchange",
          "collection_produced",
          "other",
          "unknown"
        ],
        "title": "DirectOriginKind",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlantCreate": {
        "additionalProperties": false,
        "properties": {
          "botanical_identity_id": {
            "format": "uuid",
            "title": "Botanical Identity Id",
            "type": "string"
          },
          "collection_entry_date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartialDate"
              },
              {
                "type": "null"
              }
            ]
          },
          "direct_origin_detail": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direct Origin Detail"
          },
          "direct_origin_kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DirectOriginKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/PlantLifecycle",
            "default": "active"
          },
          "location_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "material_provenance_place_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Provenance Place Id"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "originating_sowing_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Originating Sowing Id"
          },
          "supplier_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          }
        },
        "required": [
          "botanical_identity_id"
        ],
        "title": "PlantCreate",
        "type": "object"
      },
      "PlantGroupCreate": {
        "additionalProperties": false,
        "properties": {
          "botanical_identity_id": {
            "format": "uuid",
            "title": "Botanical Identity Id",
            "type": "string"
          },
          "collection_entry_date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartialDate"
              },
              {
                "type": "null"
              }
            ]
          },
          "direct_origin_detail": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direct Origin Detail"
          },
          "direct_origin_kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DirectOriginKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/PlantGroupLifecycle",
            "default": "active"
          },
          "location_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "material_provenance_place_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Provenance Place Id"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "originating_sowing_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Originating Sowing Id"
          },
          "quantity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlantGroupQuantity"
              },
              {
                "type": "null"
              }
            ]
          },
          "supplier_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          }
        },
        "required": [
          "botanical_identity_id"
        ],
        "title": "PlantGroupCreate",
        "type": "object"
      },
...
      "PlantGroupResponse": {
        "properties": {
          "botanical_identity": {
            "$ref": "#/components/schemas/BotanicalIdentitySummary"
          },
          "botanical_identity_id": {
            "format": "uuid",
            "title": "Botanical Identity Id",
            "type": "string"
          },
          "collection_entry_date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartialDate"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "direct_origin_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direct Origin Detail"
          },
          "direct_origin_kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DirectOriginKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/PlantGroupLifecycle"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocationSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "location_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "material_provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeographicPlaceSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "material_provenance_place_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Provenance Place Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "originating_sowing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OriginatingSowingSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "originating_sowing_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Originating Sowing Id"
          },
          "quantity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlantGroupQuantity"
              },
              {
                "type": "null"
              }
            ]
          },
          "supplier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SupplierSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "supplier_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "botanical_identity_id",
          "botanical_identity",
          "originating_sowing_id",
          "originating_sowing",
          "direct_origin_kind",
          "direct_origin_detail",
          "supplier_id",
          "supplier",
          "material_provenance_place_id",
          "material_provenance",
          "label",
          "collection_entry_date",
          "location_id",
          "location",
          "notes",
          "created_at",
          "updated_at",
          "quantity",
          "lifecycle"
        ],
        "title": "PlantGroupResponse",
        "type": "object"
      },
      "PlantGroupUpdate": {
        "additionalProperties": false,
        "properties": {
          "botanical_identity_id": {
            "format": "uuid",
            "title": "Botanical Identity Id",
            "type": "string"
          },
          "collection_entry_date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartialDate"
              },
              {
                "type": "null"
              }
            ]
          },
          "direct_origin_detail": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direct Origin Detail"
          },
          "direct_origin_kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DirectOriginKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/PlantGroupLifecycle",
            "default": "active"
          },
          "location_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "material_provenance_place_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Provenance Place Id"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "originating_sowing_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Originating Sowing Id"
          },
          "quantity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlantGroupQuantity"
              },
              {
                "type": "null"
              }
            ]
          },
          "supplier_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          }
        },
        "required": [
          "botanical_identity_id"
        ],
        "title": "PlantGroupUpdate",
        "type": "object"
      },
...
      "PlantResponse": {
        "properties": {
          "botanical_identity": {
            "$ref": "#/components/schemas/BotanicalIdentitySummary"
          },
          "botanical_identity_id": {
            "format": "uuid",
            "title": "Botanical Identity Id",
            "type": "string"
          },
          "collection_entry_date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartialDate"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "direct_origin_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direct Origin Detail"
          },
          "direct_origin_kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DirectOriginKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/PlantLifecycle"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocationSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "location_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "material_provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeographicPlaceSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "material_provenance_place_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Provenance Place Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "originating_sowing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OriginatingSowingSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "originating_sowing_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Originating Sowing Id"
          },
          "supplier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SupplierSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "supplier_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "botanical_identity_id",
          "botanical_identity",
          "originating_sowing_id",
          "originating_sowing",
          "direct_origin_kind",
          "direct_origin_detail",
          "supplier_id",
          "supplier",
          "material_provenance_place_id",
          "material_provenance",
          "label",
          "collection_entry_date",
          "location_id",
          "location",
          "notes",
          "created_at",
          "updated_at",
          "lifecycle"
        ],
        "title": "PlantResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Florabase API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/plant-groups": {
      "get": {
        "operationId": "listPlantGroups",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlantGroupResponse"
                  },
                  "title": "Response Listplantgroups",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List All Plant Groups",
        "tags": [
          "plant-groups"
        ]
      },
      "post": {
        "operationId": "createPlantGroup",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlantGroupCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantGroupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create One Plant Group",
        "tags": [
          "plant-groups"
        ]
      }
    },
    "/api/v1/plant-groups/{plant_group_id}": {
      "get": {
        "operationId": "getPlantGroup",
        "parameters": [
          {
            "in": "path",
            "name": "plant_group_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plant Group Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantGroupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Plant Group",
        "tags": [
          "plant-groups"
        ]
      },
      "put": {
        "operationId": "updatePlantGroup",
        "parameters": [
          {
            "in": "path",
            "name": "plant_group_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plant Group Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlantGroupUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantGroupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update One Plant Group",
        "tags": [
          "plant-groups"
        ]
      }
    },
    "/api/v1/plants": {
      "get": {
        "operationId": "listPlants",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlantResponse"
                  },
                  "title": "Response Listplants",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List All Plants",
        "tags": [
          "plants"
        ]
      },
      "post": {
        "operationId": "createPlant",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlantCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create One Plant",
        "tags": [
          "plants"
        ]
      }
    },
    "/api/v1/plants/{plant_id}": {
      "get": {
        "operationId": "getPlant",
        "parameters": [
          {
            "in": "path",
            "name": "plant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Plant",
        "tags": [
          "plants"
        ]
      },
      "put": {
        "operationId": "updatePlant",
        "parameters": [
          {
            "in": "path",
            "name": "plant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlantUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update One Plant",
        "tags": [
          "plants"
        ]
      }
    },
...
  }
}